Auto group assigning on subscription, issue #3

+++ application/classes/Model/Course.php
@@ -25,9 +25,11 @@ class Model_Course extends ORM {
     'letters' => array(
       'model' => 'Letter'
     ),
+  );
+
+  protected $_belongs_to = array(
     'group' => array(
       'model' => 'Group',
-      'through' => 'courses_groups'
     ),
   );
 
@@ -148,16 +150,19 @@ class Model_Course extends ORM {
   public function delete()
   {
     $letter_ids = $this->get_letter_ids($this->id);
-    $query = DB::delete('tasks');
-    if (is_array($letter_ids))
+    if (!empty($letter_ids))
     {
-      $query->where('letter_id', 'IN', $letter_ids);
+      $query = DB::delete('tasks');
+      if (is_array($letter_ids))
+      {
+        $query->where('letter_id', 'IN', $letter_ids);
+      }
+      else
+      {
+        $query->where('letter_id', '=', $letter_ids);
+      }
+      $query->execute();
     }
-    else
-    {
-      $query->where('letter_id', '=', $letter_ids);
-    }
-    $query->execute();
     DB::delete('letters')
       ->where('course_id', '=', $this->id)
       ->execute();
